Cleanup build of wrench
Move build logic to make and make it cleaner.
