+++ main.go
@@ -1,14 +1,18 @@
 package main
 
 import (
+	"fmt"
+	"strings"
+
 	"github.com/spf13/cobra"
 	"github.com/tomologic/wrench/build"
 	"github.com/tomologic/wrench/bump"
 	"github.com/tomologic/wrench/config"
 	"github.com/tomologic/wrench/push"
 	"github.com/tomologic/wrench/run"
-	"github.com/tomologic/wrench/version"
 )
+
+var VERSION = "0.0.0"
 
 func main() {
 	var rootCmd = &cobra.Command{Use: "wrench"}
@@ -18,7 +22,16 @@ func main() {
 	push.AddToWrench(rootCmd)
 	config.AddToWrench(rootCmd)
 	run.AddToWrench(rootCmd)
-	version.AddToWrench(rootCmd)
+
+	var cmdVersion = &cobra.Command{
+		Use:   "version",
+		Short: "Version of wrench",
+		Long:  `Version of wrench`,
+		Run: func(cmd *cobra.Command, args []string) {
+			fmt.Println(strings.Trim(VERSION, "'"))
+		},
+	}
+	rootCmd.AddCommand(cmdVersion)
 
 	rootCmd.Execute()
 }
